+++ pydbtools/utils.py
@@ -12,8 +12,7 @@ import awswrangler as wr
 bucket = "mojap-athena-query-dump"
 temp_database_name_prefix = "mojap_de_temp_"
 aws_default_region = os.getenv(
-    "AWS_DEFAULT_REGION",
-    os.getenv("AWS_REGION", "eu-west-1")
+    "AWS_DEFAULT_REGION", os.getenv("AWS_REGION", "eu-west-1")
 )